Add files via upload
Basically added a few mashes (some from a former project) to see how it would look, we could use the mats/rugs from it possibly. Made some changes to Kim's whiteboxed scene but saved it separately from his as my own scene, so it should still be intact as it was.